requirements.txt
selenium==4.27.1
undetected-chromedriver==3.5.5
python-dateutil==2.8.0
tzdata==2024.2
